Added rotation to collision boxes.

+++ engine/src/level2D.rs
@@ -24,7 +24,7 @@ pub struct LevelInstance {
 pub struct ObjectInstance {
     object_id: u32,
     position: Vec2,
-    rotation: f64
+    rotation: f32
 }
 
 #[derive(Serialize, Deserialize)]
